action_recognition/ctrgcn/requirements.txt
numpy
pyyaml
torch 